// src/prisma/contract.d.ts
export type CodecTypes = PgTypes;
// ...
export type FieldInputTypes = {
  readonly public: {
    readonly Post: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'] | null;
      readonly authorId: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly TestRecord: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'] | null;
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
  };
};